Atualização 29/07/2020

+++ plano-front/src/pages/plano/TurmasDCC/index.vue
@@ -19,6 +19,7 @@
           <i class="fas fa-times"></i>
         </BaseButton>
       </template>
+
       <template v-else>
         <BaseButton
           title="Adicionar"
@@ -234,18 +235,10 @@
             :hasSearchBar="true"
           >
             <template #thead-search>
-              <input
-                type="text"
-                class="form-control input-search"
+              <InputSearch
+                v-model="searchDisciplinasModal"
                 placeholder="Pesquise nome ou codigo de uma disciplina..."
-                @input="debounceInput($event, 'searchDisciplinasModal')"
               />
-              <button
-                @click="searchDisciplinasModal = ''"
-                class="btn btn-search"
-              >
-                <i class="fas fa-times"></i>
-              </button>
             </template>
             <template #thead
               ><th style="width:25px"></th>
@@ -334,15 +327,10 @@
             :hasSearchBar="true"
           >
             <template #thead-search>
-              <input
-                type="text"
-                class="form-control input-search"
+              <InputSearch
+                v-model="searchCursosModal"
                 placeholder="Pesquise nome ou codigo de um curso..."
-                @input="debounceInput($event, 'searchCursosModal')"
               />
-              <button @click="searchCursosModal = ''" class="btn btn-search">
-                <i class="fas fa-times"></i>
-              </button>
             </template>
             <template #thead>
               <th style="width:25px"></th>
@@ -483,54 +471,26 @@
       </template>
     </BaseModal>
 
-    <!-- MODAL DELETAR -->
-    <BaseModal
+    <ModalDelete
       ref="modalDelete"
-      :modalOptions="{
-        title: 'Deletar turma',
-        position: 'center',
-        hasBackground: true,
-        hasFooter: true,
-      }"
-      :customStyles="'width:450px; font-size:14px'"
+      :isDeleting="!!Deletar.length"
+      @btn-deletar="deleteSelectedTurmas"
     >
-      <template #modal-body>
-        <p class="w-100 m-0">
-          Tem certeza que deseja deletar a(s) turma(s) selecionadas?
-        </p>
-        <ul class="list-group list-deletar w-100 mt-2">
-          <li
-            v-for="turma in Deletar"
-            class="list-group-item"
-            :key="'delete' + turma.id"
-          >
-            <span class="mr-1"> <b> Semestre: </b>{{ turma.periodo }} </span>
-            <span class="mr-1">
-              <b> Disciplina: </b>{{ turma.disciplina.nome }}
-            </span>
-            <span class="mr-1"><b> Turma: </b> {{ turma.letra }} </span>
-          </li>
-        </ul>
-      </template>
-      <template #modal-footer>
-        <BaseButton
-          class="paddingX-20"
-          :type="'text'"
-          :color="'gray'"
-          @click="closeModalDelete()"
-        >
-          Cancelar
-        </BaseButton>
-        <BaseButton
-          class="paddingX-20 ml-auto"
-          :type="'text'"
-          :color="'red'"
-          @click="deleteSelectedTurma()"
-        >
-          Deletar
-        </BaseButton>
-      </template>
-    </BaseModal>
+      <li v-if="!Deletar.length" class="list-group-item">
+        Nenhuma turma selecionada.
+      </li>
+      <li
+        v-for="turma in Deletar"
+        class="list-group-item"
+        :key="'delete' + turma.id"
+      >
+        <span>
+          <b> Semestre: </b>{{ turma.periodo }} - <b>Turma: </b>
+          {{ turma.letra }}
+        </span>
+        <span> <b> Disciplina: </b>{{ turma.disciplina.nome }} </span>
+      </li>
+    </ModalDelete>
 
     <!-- MODAL AJUDA -->
     <BaseModal
@@ -609,16 +569,17 @@ import {
   toggleOrdination,
   toggleItemInArray,
   notification,
-  debounceInput,
   tableLoading,
 } from "@/common/mixins";
 import {
+  ModalDelete,
   PageHeader,
   BaseTable,
   BaseModal,
   NavTab,
   BodyModalEditTurma,
   BaseButton,
+  InputSearch,
 } from "@/components/ui";
 import { normalizeText } from "@/common/utils";
 import NovaTurmaRow from "./NovaTurmaRow.vue";
@@ -626,14 +587,9 @@ import TurmaRow from "./TurmaRow.vue";
 
 export default {
   name: "TurmasDCC",
-  mixins: [
-    toggleOrdination,
-    toggleItemInArray,
-    notification,
-    debounceInput,
-    tableLoading,
-  ],
+  mixins: [toggleOrdination, toggleItemInArray, notification, tableLoading],
   components: {
+    ModalDelete,
     TurmaRow,
     NovaTurmaRow,
     PageHeader,
@@ -642,6 +598,7 @@ export default {
     BaseTable,
     BaseModal,
     BaseButton,
+    InputSearch,
   },
   data() {
     return {
@@ -736,15 +693,8 @@ export default {
     }
   },
   methods: {
-    ...mapActions(["setLoadingState"]),
-    popoverCursoContent(curso) {
-      const { semestreInicial, alunosEntrada, alunosEntrada2 } = curso;
+    ...mapActions(["setLoadingState", "clearDelete"]),
 
-      if (semestreInicial == 1) return `<b>1º</b> - ${alunosEntrada}`;
-      else if (semestreInicial == 2) return `<b>2º</b> - ${alunosEntrada2}`;
-      else
-        return `<b>1º</b> - ${alunosEntrada} <br/> <b>2º</b> - ${alunosEntrada2}`;
-    },
     openAsideModal(modalName) {
       if (modalName === "filtros") {
         this.$refs.modalFiltros.toggle();
@@ -770,19 +720,8 @@ export default {
     selectCursosDCC() {
       this.filtroCursos.selecionados = [...this.CursosDCC];
     },
-    closeModalDelete() {
-      this.$refs.modalDelete.close();
-    },
     openModalDelete() {
-      if (this.Deletar.length) this.$refs.modalDelete.open();
-      else
-        this.showNotification({
-          type: "error",
-          message: "Nenhuma turma selecionada",
-        });
-    },
-    clearSearch(searchName) {
-      this[searchName] = "";
+      this.$refs.modalDelete.open();
     },
     setSemestreAtivo() {
       if (this.filtroSemestres.primeiro && !this.filtroSemestres.segundo)
@@ -792,6 +731,21 @@ export default {
       else if (this.filtroSemestres.primeiro && this.filtroSemestres.primeiro)
         this.filtroSemestres.ativo = 3;
       else this.filtroSemestres.ativo = undefined;
+    },
+    popoverCursoContent(curso) {
+      const { semestreInicial, alunosEntrada, alunosEntrada2 } = curso;
+
+      if (semestreInicial == 1) return `<b>1º</b> - ${alunosEntrada}`;
+      else if (semestreInicial == 2) return `<b>2º</b> - ${alunosEntrada2}`;
+      else
+        return `<b>1º</b> - ${alunosEntrada} <br/> <b>2º</b> - ${alunosEntrada2}`;
+    },
+    toggleIsAdding() {
+      this.isAdding = !this.isAdding;
+    },
+    nameIsBig(nome) {
+      if (nome.length > 4) return true;
+      else return false;
     },
     async xlsx() {
       try {
@@ -818,14 +772,17 @@ export default {
         this.setLoadingState("completed");
       }
     },
-    async deleteSelectedTurma() {
+    async deleteSelectedTurmas() {
+      if (!this.Deletar.length) return;
+
       try {
         this.setLoadingState("partial");
 
         for (let i = 0; i < this.Deletar.length; i++) {
           await turmaService.delete(this.Deletar[i].id);
         }
-        this.$store.commit("emptyDelete");
+        this.$refs.modalDelete.close();
+        this.clearDelete();
         this.showNotification({
           type: "success",
           message: "Turma(s) excluída(s).",
@@ -833,19 +790,11 @@ export default {
       } catch (error) {
         this.showNotification({
           type: "error",
-          message: error,
+          message: "Erro ao excluir turma(s).",
         });
       } finally {
         this.setLoadingState("completed");
-        this.closeModalDelete();
       }
-    },
-    toggleIsAdding() {
-      this.isAdding = !this.isAdding;
-    },
-    nameIsBig(nome) {
-      if (nome.length > 4) return true;
-      else return false;
     },
   },
   computed: {
